requirements.txt
gunicorn==20.0.4
python-decouple==3.3
django-cors-headers==3.7.0
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
psycopg2
markdown
django-filter
whitenoise
